Move loss utils in the modeling file
